Fix job summary stats issue
